requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.5
fitz==0.0.1.dev2
langchain==0.3.7
langchain_chroma==0.1.4
langchain_community==0.3.7
langchain_core==0.3.19
langchain_openai==0.2.9
langchain_text_splitters==0.3.2
langgraph==0.2.53
langserve==0.3.0
pydantic==2.10.1
PyPDF2==3.0.1
python-dotenv==1.0.1
typing_extensions==4.12.2
uvicorn==0.32.1
langchain_qdrant=0.2.0
